requirements.txt
streamlit
pandas
datetime
numpy
PIL
plotly.express
plotly.graph

